Dynamic table ID allocation
Allow omitting vrfTable field and let hostvrf-cni to allocate the
available table ID instead. The algorithm for finding the free table is
pretty simple. We'll start from table ID N and if there's any existing
route on the table, check the N+1. This is O(N) in the worst case where
N is the number of VRFs, but we don't expect thousands or millions of
VRFs, so shouldn't be a big problem.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>