Add protocol ID to routes
On the host network namespace, add the protocol number to the routes
installed by hostvrf-cni. This helps other softwares (e.g. FRR) to
manage the route ownership. Also introduce a knob to adjust the protocol
number in case of the collision.

Note that we won't rely on this protocol number to perform GC for
example, because it is possible that multiple hostvrf-cnis are running.
We need to update/remove the routes based on the prefix in the end.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>