Introduce Isolation Mode
Introduce a knob to control "Isolation Mode". Isolation Mode specifies
what kind of the isolation the hostvrf-cni provides. Currently, two
isolation levels are defined.

none: No isolation between default VRF. The traffic going out from
      containers can fall through to the main routing table (no
      unreachable default route inserted to the VRF table).
      Container network is reachable from the default VRF.

full: The current behavior. The unreachable default route will be
      inserted, so the traffic going out from containers cannot fall
      through to the main routing table. Container network is
      unreachable from the default VRF.

In the future, we may want to add more fine-grained isolation level. For
the sake of extensibilty, it is defined as a string field.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>