Implement egress masquerading
Implement egress masquerading similar to the bridge CNI's ipMasq.
Currently, we support two modes. "disabled" which disables the egress
NAT entirely and "hostip" that masquerades source IP address to the
host's IP address.

"hostip" mode works even if there's overlapping IPs among multiple VRFs
on the same host. This is accomplished by distinguishing the same 5
tuples using conntrack zone and encoding VRF's table ID to the conntrack
mark in the request path. In the reply path, the host IP + Port number
uniquely identifies the conntrack entry (assuming there's no host IP
overlapping) and the conntrack mark identifies the table to lookup. The
actual lookup is realized by the special IP rule matches to the mark.

Note that because we start to encode table ID to the conntrack zone, we
need to limit the table ID to 16bit.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>