Make routing isolation configurable
Make the installation of the unreachable default route configurable. We
also change the default to be not installed so that it just works with
the default settings without any route leaking settings (the
"fallthrough" mode). Advanced users can set it to "isolated" and do
route leaking using FRR for example.

Signed-off-by: Yutaro Hayakawa <yhayakawa3720@gmail.com>